+++ src/compiler/scala/tools/nsc/typechecker/Namers.scala
@@ -1600,12 +1600,12 @@ trait Namers extends MethodSynthesis {
       import SymValidateErrors._
       def fail(kind: SymValidateErrors.Value) = SymbolValidationError(sym, kind)
 
-      def checkWithDeferred(flag: Int) {
+      def checkWithDeferred(flag: Long) {
         if (sym hasFlag flag)
           AbstractMemberWithModiferError(sym, flag)
       }
-      def checkNoConflict(flag1: Int, flag2: Int) {
-        if (sym hasAllFlags flag1.toLong | flag2)
+      def checkNoConflict(flag1: Long, flag2: Long) {
+        if (sym hasAllFlags flag1 | flag2)
           IllegalModifierCombination(sym, flag1, flag2)
       }
       if (sym.isImplicit) {
